+++ spec/unit/installer/user_project_integrator_spec.rb
@@ -21,10 +21,12 @@ module Pod
         @target.client_root = sample_project_path.dirname
         @target.user_project = Xcodeproj::Project.open(@sample_project_path)
         @target.user_target_uuids = ['A346496C14F9BE9A0080D870']
+        @target.user_build_configurations = { 'Release' => :release, 'Debug' => :debug }
         @empty_library = AggregateTarget.new(@podfile.target_definitions[:empty], config.sandbox)
         @empty_library.client_root = sample_project_path.dirname
         @empty_library.user_project = @target.user_project
         @empty_library.user_target_uuids = ['C0C495321B9E5C47004F9854']
+        @empty_library.user_build_configurations = { 'Release' => :release, 'Debug' => :debug }
         @integrator = UserProjectIntegrator.new(@podfile, config.sandbox, temporary_directory, [@target, @empty_library])
       end